requirements_mobile.txt
flet
yt-dlp
Pillow
requests
